requirements.txt
psycopg[binary]
psycopg_pool
python-dotenv